+++ src/panel/main/Update.vue
@@ -1,5 +1,5 @@
 <template>
-  <Popup :buttons="[{ name: '更新', shouldEmphasize: true, onClick: onConfirm }, { name: '取消' }]">
+  <Popup width="wide" :buttons="[{ name: '更新', shouldEmphasize: true, onClick: onConfirm }, { name: '取消' }]">
     <!-- eslint-disable-next-line vue/no-v-html -->
     <div class="p-1.5 text-left" v-html="changelog" />
   </Popup>